+++ src/Controller/Admin/SystemsController.php
@@ -23,12 +23,11 @@
 namespace MeCms\Controller\Admin;
 
 use Cake\Core\Configure;
-use Cake\Filesystem\Folder;
 use Cake\Network\Exception\InternalErrorException;
 use Cake\Routing\Router;
 use MeCms\Controller\AppController;
+use MeCms\Core\Plugin;
 use MeTools\Cache\Cache;
-use MeTools\Core\Plugin;
 use MeTools\Utility\Apache;
 
 /**
@@ -44,8 +43,9 @@ class SystemsController extends AppController {
 	 */
 	public function isAuthorized($user = NULL) {		
 		//Only admins can clear all temporary files or logs
-		if($this->request->isAction('tmp_cleaner') && in_array($this->request->param('pass.0'), ['all', 'logs']))
+		if($this->request->isAction('tmp_cleaner') && in_array($this->request->param('pass.0'), ['all', 'logs'])) {
 			return $this->Auth->isGroup('admin');
+        }
 		
 		//Admins and managers can access other actions
 		return $this->Auth->isGroup(['admin', 'manager']);
@@ -65,19 +65,22 @@ class SystemsController extends AppController {
 		$this->loadComponent('MeCms.KcFinder');
 		
 		//Checks for KCFinder
-		if(!$this->KcFinder->checkKcfinder())
+		if(!$this->KcFinder->checkKcfinder()) {
 			throw new InternalErrorException(__d('me_cms', '{0} is not present into {1}', 'KCFinder', rtr($this->KcFinder->getKcfinderPath())));
-		
+        }
+        
 		//Checks for the files directory (`APP/webroot/files`)
-		if(!$this->KcFinder->checkFiles())
+		if(!$this->KcFinder->checkFiles()) {
 			throw new InternalErrorException(__d('me_tools', 'File or directory {0} not writeable', rtr($this->KcFinder->getFilesPath())));
+        }
 		
 		//Gets the supperted types from configuration
 		$types = $this->KcFinder->getTypes();
 		
 		//If there's only one type, it automatically sets the query value
-		if(!$this->request->query('type') && count($types) < 2)
+		if(!$this->request->query('type') && count($types) < 2) {
 			$this->request->query['type'] = fk($types);
+        }
 		
 		//Gets the type from the query and the types from configuration
 		$type = $this->request->query('type');
@@ -85,23 +88,22 @@ class SystemsController extends AppController {
 		$locale = substr(\Cake\I18n\I18n::locale(), 0, 2);
 		
 		//Checks the type, then sets the KCFinder path
-		if($type && array_key_exists($type, $types))
+		if($type && array_key_exists($type, $types)) {
 			$this->set('kcfinder', sprintf('%s/kcfinder/browse.php?lang=%s&type=%s', Router::url('/vendor', TRUE), empty($locale) ? 'en' : $locale, $type));
+        }
 		
 		$this->set('types', array_combine(array_keys($types), array_keys($types)));
 	}
 
 	/**
 	 * Changelogs viewer
+     * @uses MeCms\Core\Plugin:all()
+     * @uses MeCms\Core\Plugin:path()
 	 */
 	public function changelogs() {
-		//Gets all changelog files. 
-		//Searchs into `ROOT` and all loaded plugins.
-		foreach(am([ROOT.DS], Plugin::path()) as $path) {
-			//For each changelog file in the current path
-			foreach((new Folder($path))->find('CHANGELOG(\..+)?') as $file)
-				$files[] = rtr($path.$file);
-		}
+        foreach(Plugin::all() as $plugin) {
+            $files[$plugin] = rtr(Plugin::path($plugin, 'CHANGELOG.md', TRUE));
+        }
 		
 		//If a changelog file has been specified
 		if($this->request->query('file') && $this->request->is('get')) {
@@ -116,69 +118,103 @@ class SystemsController extends AppController {
 	
 	/**
 	 * System checkup
-	 * @uses MeTools\Core\Plugin::version()
-	 * @uses MeTools\Core\Plugin::versions()
+	 * @uses MeCms\Core\Plugin::all()
+	 * @uses MeCms\Core\Plugin::path()
 	 * @uses MeTools\Utility\Apache::module()
 	 * @uses MeTools\Utility\Apache::version()
 	 */
 	public function checkup() {
         $checkup['apache'] = [
-            'expires'	=> Apache::module('mod_expires'),
-            'rewrite'	=> Apache::module('mod_rewrite'),
-            'version'	=> Apache::version(),
+            'expires' => Apache::module('mod_expires'),
+            'rewrite' => Apache::module('mod_rewrite'),
+            'version' => Apache::version(),
         ];
         
         $checkup['backups'] = [
-            'path'		=> rtr(BACKUPS),
+            'path' => rtr(BACKUPS),
             'writeable'	=> folder_is_writable(BACKUPS),
         ];
         
         $checkup['cache'] = Cache::enabled();
         
         $checkup['executables'] = [
-            'clean-css'		=> which('cleancss'),
-            'UglifyJS 2'	=> which('uglifyjs'),
+            'clean-css' => which('cleancss'),
+            'UglifyJS 2' => which('uglifyjs'),
         ];
         
         //Checks for PHP's extensions
-        foreach(['exif', 'imagick', 'mcrypt', 'zip'] as $extension)
+        foreach(['exif', 'imagick', 'mcrypt', 'zip'] as $extension) {
             $checkup['php_extensions'][$extension] = extension_loaded($extension);
+        }
         
         $checkup['plugins'] = [
-            'cakephp_version'	=> Configure::version(),
-            'plugins_version'	=> Plugin::versions('MeCms'),
-            'mecms_version'		=> Plugin::version('MeCms'),
+            'cakephp' => Configure::version(),
+            'mecms'	=> trim(file_get_contents(Plugin::path(MECMS, 'version'))),
         ];
         
+        //Gets plugins versions
+        foreach(Plugin::all(['exclude' => MECMS]) as $plugin) {
+            $file = Plugin::path(MECMS, 'version', TRUE);
+            
+            if($file) {
+                $checkup['plugins']['plugins'][$plugin] = trim(file_get_contents($file));
+            }
+        }
+        
         //Checks for temporary directories
-        foreach([CACHE, LOGS, THUMBS, TMP] as $path)
+        foreach([CACHE, LOGS, THUMBS, TMP] as $path) {
             $checkup['temporary'][] = ['path' => rtr($path), 'writeable' => folder_is_writable($path)];
+        }
         
         //Checks for webroot directories
-        foreach([ASSETS, BANNERS, PHOTOS, WWW_ROOT.'files', WWW_ROOT.'fonts'] as $path)
+        foreach([ASSETS, BANNERS, PHOTOS, WWW_ROOT.'files', WWW_ROOT.'fonts'] as $path) {
             $checkup['webroot'][] = ['path' => rtr($path), 'writeable' => folder_is_writable($path)];
+        }
         
         array_walk($checkup, function($value, $key) {
             $this->set($key, $value);
         });
 	}
-	
-	/**
-	 * Temporary cleaner (assets, cache, logs and thumbnails)
+    
+    /**
+     * Internal function to clear the cache
+     * @return bool
+	 * @uses MeTools\Cache\Cache::clearAll()
+     */
+    protected function clear_cache() {
+        return !array_search(FALSE, Cache::clearAll(), TRUE);
+    }
+
+    /**
+     * Internal function to clear the sitemap
+     * @return bool
+     */
+    protected function clear_sitemap() {
+        if(!is_readable(SITEMAP)) {
+            return TRUE;
+        }
+        
+        return (new \Cake\Filesystem\File(SITEMAP))->delete();
+    }
+
+    /**
+	 * Temporary cleaner (assets, cache, logs, sitemap and thumbnails)
 	 * @param string $type Type
      * @throws InternalErrorException
-	 * @uses MeTools\Cache\Cache::clearAll()
+     * @uses clear_cache()
+     * @uses clear_sitemap()
 	 */
 	public function tmp_cleaner($type) {
-		if(!$this->request->is(['post', 'delete']))
+		if(!$this->request->is(['post', 'delete'])) {
 			return $this->redirect(['action' => 'tmp_viewer']);
+        }
 		
 		switch($type) {
 			case 'all':
-				$success = clear_dir(ASSETS) && clear_dir(LOGS) && Cache::clearAll() && clear_dir(THUMBS);
+				$success = clear_dir(ASSETS) && clear_dir(LOGS) && self::clear_cache() && self::clear_sitemap() && clear_dir(THUMBS);
 				break;
 			case 'cache':
-				$success = Cache::clearAll();
+				$success = self::clear_cache();
 				break;
 			case 'assets':
 				$success = clear_dir(ASSETS);
@@ -186,6 +222,9 @@ class SystemsController extends AppController {
 			case 'logs':
 				$success = clear_dir(LOGS);
 				break;
+            case 'sitemap':
+                $success = self::clear_sitemap();
+                break;
 			case 'thumbs':
 				$success = clear_dir(THUMBS);
 				break;
@@ -193,25 +232,30 @@ class SystemsController extends AppController {
                 throw new InternalErrorException(__d('me_cms', 'Unknown command type'));
 		}
 		
-		if(!empty($success))
+		if(!empty($success)) {
 			$this->Flash->success(__d('me_cms', 'The operation has been performed correctly'));
-		else
+        }
+		else {
 			$this->Flash->error(__d('me_cms', 'The operation has not been performed correctly'));
+        }
 		
 		return $this->redirect(['action' => 'tmp_viewer']);
 	}
 	
 	/**
-	 * Temporary viewer (assets, cache, logs and thumbnails)
+	 * Temporary viewer (assets, cache, logs, sitemap and thumbnails)
 	 */
 	public function tmp_viewer() {
+        $sitemap = is_readable(SITEMAP) ? filesize(SITEMAP) : 0;
+        
         $this->set([
-			'cache_size'	=> dirsize(CACHE),
-			'cache_status'	=> Cache::enabled(),
-			'assets_size'	=> dirsize(ASSETS),
-			'logs_size'     => dirsize(LOGS),
-			'thumbs_size'	=> dirsize(THUMBS),
-			'total_size'	=> dirsize(CACHE) + dirsize(ASSETS) + dirsize(LOGS) + dirsize(THUMBS),
+			'cache_size' => dirsize(CACHE),
+			'cache_status' => Cache::enabled(),
+			'assets_size' => dirsize(ASSETS),
+			'logs_size' => dirsize(LOGS),
+            'sitemap_size' => $sitemap,
+			'thumbs_size' => dirsize(THUMBS),
+			'total_size' => dirsize(CACHE) + dirsize(ASSETS) + dirsize(LOGS) + $sitemap + dirsize(THUMBS),
         ]);
 	}
 }