+++ src/Shell/UpdateShell.php
@@ -22,74 +22,133 @@
  */
 namespace MeCms\Shell;
 
-use MeTools\Console\Shell;
+use MeCms\Shell\BaseUpdateShell;
 
 /**
  * Applies updates
  */
-class UpdateShell extends Shell {
+class UpdateShell extends BaseUpdateShell {
 	/**
-	 * Database connection
-	 * @see initialize()
-	 * @var resource 
+	 * Updates to 2.6.0 version
+	 * @uses MeCms\Shell\BaseUpdateShell::$connection
+     * @uses MeCms\Shell\BaseUpdateShell::_checkColumn()
 	 */
-	protected $connection;
-	
-	/**
-	 * Now for MySql
-	 * @see initialize()
-	 * @var string 
-	 */
-	protected $now;
-
-	/**
-	 * Initialize
-	 * @uses $connection
-	 */
-	public function initialize() {
-        parent::initialize();
-		
-		//Gets database connection
-		$this->connection = \Cake\Datasource\ConnectionManager::get('default');
-		
-		//Sets now for MySql
-		$this->now = (new \Cake\I18n\Time)->now()->i18nFormat(FORMAT_FOR_MYSQL);
-	}
+	public function to2v6v0() {
+		$this->loadModel('MeCms.BannersPositions');
+		$this->loadModel('MeCms.PhotosAlbums');
+		$this->loadModel('MeCms.PostsCategories');
+		$this->loadModel('MeCms.Tags');
+		$this->loadModel('MeCms.UsersGroups');
+        
+        //Adds "created" field to the banners positions table and sets the default value
+        if(!$this->_checkColumn('created', $this->BannersPositions->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `created` DATETIME NULL AFTER `banner_count`;', $this->BannersPositions->table()));
+            $this->BannersPositions->query()->update()->set(['created' => $this->now])->execute();
+        }
+        
+        //Adds "modified" field to the banners positions table and sets the default value
+        if(!$this->_checkColumn('modified', $this->BannersPositions->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `modified` DATETIME NULL AFTER `created`;', $this->BannersPositions->table()));
+            $this->BannersPositions->query()->update()->set(['modified' => $this->now])->execute();
+        }
+        
+        //Adds "created" field to the photos albums table and sets the default value
+        if(!$this->_checkColumn('created', $this->PhotosAlbums->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `created` DATETIME NULL AFTER `photo_count`;', $this->PhotosAlbums->table()));
+            $this->PhotosAlbums->query()->update()->set(['created' => $this->now])->execute();
+        }
+        
+        //Adds "modified" field to the photos albums table and sets the default value
+        if(!$this->_checkColumn('modified', $this->PhotosAlbums->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `modified` DATETIME NULL AFTER `created`;', $this->PhotosAlbums->table()));
+            $this->PhotosAlbums->query()->update()->set(['modified' => $this->now])->execute();
+        }
+        
+        //Adds "created" field to the posts categories table and sets the default value
+        if(!$this->_checkColumn('created', $this->PostsCategories->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `created` DATETIME NULL AFTER `post_count`;', $this->PostsCategories->table()));
+            $this->PostsCategories->query()->update()->set(['created' => $this->now])->execute();
+        }
+        
+        //Adds "modified" field to the posts categories table and sets the default value
+        if(!$this->_checkColumn('modified', $this->PostsCategories->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `modified` DATETIME NULL AFTER `created`;', $this->PostsCategories->table()));
+            $this->PostsCategories->query()->update()->set(['modified' => $this->now])->execute();
+        }
+        
+        //Adds "created" field to the tags table and sets the default value
+        if(!$this->_checkColumn('created', $this->Tags->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `created` DATETIME NULL AFTER `post_count`;', $this->Tags->table()));
+            $this->Tags->query()->update()->set(['created' => $this->now])->execute();
+        }
+        
+        //Adds "modified" field to the tags table and sets the default value
+        if(!$this->_checkColumn('modified', $this->Tags->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `modified` DATETIME NULL AFTER `created`;', $this->Tags->table()));
+            $this->Tags->query()->update()->set(['modified' => $this->now])->execute();
+        }
+        
+        //Adds "created" field to the users groups table and sets the default value
+        if(!$this->_checkColumn('created', $this->UsersGroups->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `created` DATETIME NULL AFTER `user_count`;', $this->UsersGroups->table()));
+            $this->UsersGroups->query()->update()->set(['created' => $this->now])->execute();
+        }
+        
+        //Adds "modified" field to the users groups table and sets the default value
+        if(!$this->_checkColumn('modified', $this->UsersGroups->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `modified` DATETIME NULL AFTER `created`;', $this->UsersGroups->table()));
+            $this->UsersGroups->query()->update()->set(['modified' => $this->now])->execute();
+        }
+    }
+    
 	/**
 	 * Updates to 2.2.1 version
-	 * @uses $connection
+	 * @uses MeCms\Shell\BaseUpdateShell::$connection
 	 */
 	public function to2v2v1() {
 		$this->loadModel('MeCms.Tags');
 		
 		//For each tag, it replaces the hyphen with space
-		foreach($this->Tags->find()->where(['tag LIKE' => '%-%'])->toArray() as $tag)
+		foreach($this->Tags->find()->where(['tag LIKE' => '%-%'])->toArray() as $tag) {
 			$this->Tags->query()->update()
 				->set(['tag' => str_replace('-', ' ', $tag->tag)])
 				->where(['id' => $tag->id])
 				->execute();
+        }
 	}
 	
 	/**
 	 * Updates to 2.1.9 version
-	 * @uses $connection
+	 * @uses MeCms\Shell\BaseUpdateShell::$connection
+     * @uses MeCms\Shell\BaseUpdateShell::_checkColumn()
 	 */
 	public function to2v1v9() {
 		$this->loadModel('MeCms.Banners');
 		$this->loadModel('MeCms.Photos');
 		
-		//Adds "created" and "modified" field to the banners table and sets the default value
-		$this->connection->execute(sprintf('ALTER TABLE `%s` ADD `created` DATETIME NULL AFTER `click_count`, ADD `modified` DATETIME NULL AFTER `created`;', $this->Banners->table()));
-		$this->Banners->query()->update()->set(['created' => $this->now, 'modified' => $this->now])->execute();
-		
-		//Adds "modified" field to the photos table and sets the default value
-		$this->connection->execute(sprintf('ALTER TABLE `%s` ADD `modified` DATETIME NULL AFTER `created`;', $this->Photos->table()));
-		$this->Photos->query()->update()->set(['modified' => $this->now])->execute();
+        //Adds "created" field to the banners table and sets the default value
+        if(!$this->_checkColumn('created', $this->Banners->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `created` DATETIME NULL AFTER `click_count`;', $this->Banners->table()));
+            $this->Banners->query()->update()->set(['created' => $this->now])->execute();
+        }
+        
+        //Adds "modified" field to the banners table and sets the default value
+        if(!$this->_checkColumn('modified', $this->Banners->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `modified` DATETIME NULL AFTER `created`;', $this->Banners->table()));
+            $this->Banners->query()->update()->set(['modified' => $this->now])->execute();
+        }
+        
+        //Adds "modified" field to the photos table and sets the default value
+        if(!$this->_checkColumn('modified', $this->Photos->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `modified` DATETIME NULL AFTER `created`;', $this->Photos->table()));
+            $this->Photos->query()->update()->set(['modified' => $this->now])->execute();
+        }
 	}
 	
 	/**
 	 * Updates to 2.1.8 version
-	 * @uses $connection
+	 * @uses MeCms\Shell\BaseUpdateShell::$connection
+     * @uses MeCms\Shell\BaseUpdateShell::_checkColumn()
 	 */
 	public function to2v1v8() {
 		$this->loadModel('MeCms.Photos');
@@ -99,20 +158,23 @@ class UpdateShell extends Shell {
 		$this->Tags->deleteAll(['post_count' => 0]);
 				
 		//For each tag, it replaces the hyphen with space
-		foreach($this->Tags->find()->toArray() as $tag)
+		foreach($this->Tags->find()->toArray() as $tag) {
 			$this->Tags->query()->update()
 				->set(['tag' => str_replace('-', ' ', $tag->tag)])
 				->where(['id' => $tag->id])
 				->execute();
+        }
 		
 		//Adds "created" field to the photos table and sets the default value
-		$this->connection->execute(sprintf('ALTER TABLE `%s` ADD `created` DATETIME NULL DEFAULT NULL AFTER `description`;', $this->Photos->table()));
-		$this->Photos->query()->update()->set(['created' => $this->now])->execute();
+        if(!$this->_checkColumn('created', $this->Photos->table())) {
+            $this->connection->execute(sprintf('ALTER TABLE `%s` ADD `created` DATETIME NULL DEFAULT NULL AFTER `description`;', $this->Photos->table()));
+            $this->Photos->query()->update()->set(['created' => $this->now])->execute();
+        }
 	}
 	
 	/**
 	 * Updates to 2.1.7 version
-	 * @uses $connection
+	 * @uses MeCms\Shell\BaseUpdateShell::$connection
 	 */
 	public function to2v1v7() {
 		$this->loadModel('MeCms.Tags');
@@ -123,12 +185,12 @@ class UpdateShell extends Shell {
 	/**
 	 * Gets the option parser instance and configures it.
 	 * @return ConsoleOptionParser
-	 * @uses MeTools\Shell\InstallShell::getOptionParser()
 	 */
 	public function getOptionParser() {
 		$parser = parent::getOptionParser();
 		
 		return $parser->addSubcommands([
+            'to2v6v0' => ['help' => __d('me_cms', 'Updates to {0} version', '2.6.0')],
 			'to2v2v1' => ['help' => __d('me_cms', 'Updates to {0} version', '2.2.1')],
 			'to2v1v9' => ['help' => __d('me_cms', 'Updates to {0} version', '2.1.9')],
 			'to2v1v8' => ['help' => __d('me_cms', 'Updates to {0} version', '2.1.8')],