+++ src/Utility/Sitemap.php
@@ -25,31 +25,51 @@ namespace MeCms\Utility;
 
 use Cake\ORM\TableRegistry;
 use MeCms\Utility\SitemapBuilder;
+use MeTools\Cache\Cache;
 
 /**
  * This class contains methods called by the `SitemapBuilder`.
  * Each method must be return an array or urls to add to the sitemap.
+ * 
+ * This class contains methods that will be called automatically.
+ * You do not need to call these methods manually.
  */
 class Sitemap extends SitemapBuilder {
     /**
      * Method that returns pages urls
      * @return array
-     * @uses MeCms\Utility\SitemapBuilder::url()
+     * @uses MeCms\Utility\SitemapBuilder::parse()
      */
     public static function pages() {
-        $pages = TableRegistry::get('MeCms.Pages')->find('active')->select(['slug']);
+        $table = TableRegistry::get('MeCms.Pages');
+        
+        $url = Cache::read('sitemap', $table->cache);
+        
+        if($url) {
+            return $url;
+        }
+        
+        $pages = $table->find('active')
+            ->select(['slug', 'modified']);
         
         if($pages->isEmpty()) {
             return [];
         }
         
+        $latest = $table->find('active')
+            ->select(['modified'])
+            ->order(['Pages.modified' => 'DESC'])
+            ->firstOrFail();
+        
         //Adds pages index
-        $url = [self::url(['_name' => 'pages'])];
+        $url = [self::parse(['_name' => 'pages'], ['lastmod' => $latest->modified])];
 
         //Adds all pages
         $url = am($url, array_map(function($page) {
-            return self::url(['_name' => 'page', $page->slug]);
+            return self::parse(['_name' => 'page', $page->slug], ['lastmod' => $page->modified]);
         }, $pages->toArray()));
+        
+        Cache::write('sitemap', $url, $table->cache);
         
         return $url;
     }
@@ -57,31 +77,48 @@ class Sitemap extends SitemapBuilder {
     /**
      * Method that returns photos urls
      * @return array
-     * @uses MeCms\Utility\SitemapBuilder::url()
+     * @uses MeCms\Utility\SitemapBuilder::parse()
      */
     public static function photos() {
-        $albums = TableRegistry::get('MeCms.PhotosAlbums')->find('active')
+        $table = TableRegistry::get('MeCms.PhotosAlbums');
+        
+        $url = Cache::read('sitemap', $table->cache);
+        
+        if($url) {
+            return $url;
+        }
+        
+        $albums = $table->find('active')
             ->select(['id', 'slug'])
             ->contain(['Photos' => function($q) {
-                return $q->select(['id', 'album_id']);
+                return $q
+                    ->select(['id', 'album_id', 'modified'])
+                    ->order(['Photos.modified' => 'DESC']);
             }]);
         
         if($albums->isEmpty()) {
             return [];
         }
         
+        $latest = $table->Photos->find('active')
+            ->select(['modified'])
+            ->order(['Photos.modified' => 'DESC'])
+            ->firstOrFail();
+        
         //Adds albums index
-        $url = [self::url(['_name' => 'albums'])];
+        $url = [self::parse(['_name' => 'albums'], ['lastmod' => $latest->modified])];
 
         foreach($albums->toArray() as $album) {
             //Adds the album
-            $url[] = self::url(['_name' => 'album', $album->slug]);
+            $url[] = self::parse(['_name' => 'album', $album->slug], ['lastmod' => $album->photos[0]->modified]);
 
             //Adds the photos
             $url = am($url, array_map(function($photo) use($album) {
-                return self::url(['_name' => 'photo', 'slug' => $album->slug, 'id' => $photo->id]);
+                return self::parse(['_name' => 'photo', 'slug' => $album->slug, 'id' => $photo->id], ['lastmod' => $photo->modified]);
             }, $album->photos));
         }
+        
+        Cache::write('sitemap', $url, $table->cache);
             
         return $url;
     }
@@ -89,35 +126,52 @@ class Sitemap extends SitemapBuilder {
     /**
      * Method that returns posts urls
      * @return array
-     * @uses MeCms\Utility\SitemapBuilder::url()
+     * @uses MeCms\Utility\SitemapBuilder::parse()
      */
-    public static function posts() {        
-        $categories = TableRegistry::get('MeCms.PostsCategories')->find('active')
+    public static function posts() {
+        $table = TableRegistry::get('MeCms.PostsCategories');
+        
+        $url = Cache::read('sitemap', $table->cache);
+        
+        if($url) {
+            return $url;
+        }
+        
+        $categories = $table->find('active')
             ->select(['id', 'slug'])
             ->contain(['Posts' => function($q) {
-                return $q->select(['category_id', 'slug']);
+                return $q
+                    ->select(['category_id', 'slug', 'modified'])
+                    ->order(['Posts.modified' => 'DESC']);
             }]);
         
         if($categories->isEmpty()) {
             return [];
         }
         
+        $latest = $table->Posts->find('active')
+            ->select(['modified'])
+            ->order(['Posts.modified' => 'DESC'])
+            ->firstOrFail();
+        
         //Adds posts index, categories index and posts search
         $url = [
-            self::url(['_name' => 'posts']),
-            self::url(['_name' => 'posts_categories']),
-            self::url(['_name' => 'posts_search']),
+            self::parse(['_name' => 'posts'], ['lastmod' => $latest->modified]),
+            self::parse(['_name' => 'posts_categories']),
+            self::parse(['_name' => 'posts_search'], ['priority' => '0.2']),
         ];
         
         foreach($categories as $category) {
             //Adds the category
-            $url[] = self::url(['_name' => 'posts_category', $category->slug]);
+            $url[] = self::parse(['_name' => 'posts_category', $category->slug], ['lastmod' => $category->posts[0]->modified]);
             
             //Adds the posts
             $url = am($url, array_map(function($post) {
-                return self::url(['_name' => 'post', $post->slug]);
+                return self::parse(['_name' => 'post', $post->slug], ['lastmod' => $post->modified]);
             }, $category->posts));
         }
+        
+        Cache::write('sitemap', $url, $table->cache);
         
         return $url;
     }
@@ -125,10 +179,19 @@ class Sitemap extends SitemapBuilder {
     /**
      * Method that returns posts tags urls
      * @return array
-     * @uses MeCms\Utility\SitemapBuilder::url()
+     * @uses MeCms\Utility\SitemapBuilder::parse()
      */
-    public static function posts_tags() {            
-        $tags = TableRegistry::get('MeCms.Tags')->find('all')
+    public static function posts_tags() {
+        $table = TableRegistry::get('MeCms.Tags');
+        
+        $url = Cache::read('sitemap', $table->cache);
+        
+        if($url) {
+            return $url;
+        }
+        
+        $tags = $table->find('all')
+            ->select(['tag', 'modified'])
 			->order(['tag' => 'ASC'])
 			->where(['post_count >' => 0]);
         
@@ -136,13 +199,20 @@ class Sitemap extends SitemapBuilder {
             return [];
         }
         
+        $latest = $table->find()
+            ->select(['modified'])
+            ->order(['Tags.modified' => 'DESC'])
+            ->firstOrFail();
+        
         //Adds the tags index
-        $url[] = self::url(['_name' => 'posts_tags']);
+        $url[] = self::parse(['_name' => 'posts_tags'], ['lastmod' => $latest->modified]);
 
         //Adds all tags
         $url = am($url, array_map(function($tag) {
-            return self::url(['_name' => 'posts_tag', $tag->slug]);
+            return self::parse(['_name' => 'posts_tag', $tag->slug], ['lastmod' => $tag->modified]);
         }, $tags->toArray()));
+        
+        Cache::write('sitemap', $url, $table->cache);
         
         return $url;
     }
@@ -150,7 +220,7 @@ class Sitemap extends SitemapBuilder {
     /**
      * Method that returns static pages urls
      * @return array
-     * @uses MeCms\Utility\SitemapBuilder::url()
+     * @uses MeCms\Utility\SitemapBuilder::parse()
      * @uses MeCms\Utility\StaticPage::all()
      */
     public static function static_pages() {
@@ -158,7 +228,7 @@ class Sitemap extends SitemapBuilder {
         
         //Adds static pages
         $url = array_map(function($page) {
-            return self::url(['_name' => 'page', $page->slug]);
+            return self::parse(['_name' => 'page', $page->slug], ['lastmod' => $page->modified]);
         }, $statics);
         
         return $url;
@@ -167,14 +237,14 @@ class Sitemap extends SitemapBuilder {
     /**
      * Method that returns systems urls.
      * @return array
-     * @uses MeCms\Utility\SitemapBuilder::url()
+     * @uses MeCms\Utility\SitemapBuilder::parse()
      */
     public static function systems() {
         $url = [];
         
         //Contact form
         if(config('frontend.contact_form'))
-            $url[] = self::url(['_name' => 'contact_form']);
+            $url[] = self::parse(['_name' => 'contact_form']);
         
         return $url;
     }